Move config extraction into base class
