+++ extensions/typescript-language-features/src/utils/configuration.browser.ts
@@ -2,26 +2,6 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-
-import * as vscode from 'vscode';
 import { BaseServiceConfigurationProvider } from './configuration';
 
-export class BrowserServiceConfigurationProvider extends BaseServiceConfigurationProvider {
-
-	protected extractGlobalTsdk(configuration: vscode.WorkspaceConfiguration): string | null {
-		const inspect = configuration.inspect('typescript.tsdk');
-		if (inspect && typeof inspect.globalValue === 'string') {
-			return inspect.globalValue;
-		}
-		return null;
-	}
-
-
-	protected extractLocalTsdk(configuration: vscode.WorkspaceConfiguration): string | null {
-		const inspect = configuration.inspect('typescript.tsdk');
-		if (inspect && typeof inspect.workspaceValue === 'string') {
-			return inspect.workspaceValue;
-		}
-		return null;
-	}
-}
+export class BrowserServiceConfigurationProvider extends BaseServiceConfigurationProvider { }
